feat: support dictionary masks in feature extraction and cast InternVideo encoder input to float32 for compatibility.

+++ src/temporal/internvideo_hico_module.py
@@ -140,7 +140,8 @@ class FrameEncoder(nn.Module):
 
             def forward(self, x: torch.Tensor) -> torch.Tensor:
                 batch_size, num_frames = x.shape[:2]
-                x = x.reshape(batch_size * num_frames, -1)
+                # Cast to float32 for linear layer compatibility
+                x = x.float().reshape(batch_size * num_frames, -1)
                 return self.proj(x).reshape(batch_size, num_frames, -1)
 
         return PlaceholderEncoder(self.config.hidden_dim).to(self.config.device)
